refactor: rename Result -> Response and Response module -> Responder
The previous names overlapped: ServiceCore::Result (the value object)
and ServiceCore::Response (the helpers mixin) were both about "what a
service returns", which made the public API harder to talk about.

Each layer is now named by its role:

- ServiceCore::Response (class) -- the four-key value object. Was Result.
  This matches the language used throughout the README and the blog
  post that motivates the response shape.
- ServiceCore::Responder (module) -- the mixin that provides the
  success_response / error_response / formatted_response helpers. Was
  the old ServiceCore::Response. A service includes Responder; the value
  it emits is a Response.
- ServiceCore::Output (module) -- unchanged. The internal mechanism
  layer that owns @output, dirty tracking and key validation.

service.response is added as an alias for service.output, so callers
can use either name. output is retained for backward compatibility.

Spec files are renamed in lock-step:
- spec/service_core/result_spec.rb   -> response_spec.rb
- spec/service_core/response_spec.rb -> responder_spec.rb

README and CHANGELOG are updated to use the new names throughout.

No deprecated alias is provided for the old ServiceCore::Response module
name: it was always an internal mixin consumed via `include ServiceCore`,
never documented as a public entry point. Anyone who included it
directly should switch to `include ServiceCore::Responder`.

108 examples, all green on Rails 7.2, 8.0 and 8.1; RuboCop clean.
