+++ lib/service_core/responder.rb
@@ -1,0 +1,45 @@
+require "active_support/concern"
+require "active_model/errors"
+require_relative "output"
+
+module ServiceCore
+  # Mixin that gives a service the ergonomic helpers used inside +perform+
+  # to assemble a {ServiceCore::Response}: +success_response+,
+  # +error_response+ and +formatted_response+.
+  #
+  # A class that includes +Responder+ becomes capable of responding; the
+  # value it emits is a {ServiceCore::Response}.
+  module Responder
+    extend ActiveSupport::Concern
+
+    include ServiceCore::Output
+
+    protected
+
+    def success_response(message: nil, data: nil)
+      formatted_response(status: "success", message: message, data: data)
+    end
+
+    def error_response(message:, errors: nil)
+      formatted_response(status: "error", message: message, errors: errors)
+    end
+
+    # Records the response keys on +@output+.
+    #
+    # Each key is only skipped when its value is +nil+, so legitimate
+    # falsy values (+false+, +0+, +""+) are preserved.
+    def formatted_response(status:, message: nil, data: nil, errors: nil)
+      set_output(:status, status)
+      set_output(:message, message)
+      set_output(:data, data)
+      set_output(:errors, error_messages(errors))
+      output
+    end
+
+    def error_messages(errors)
+      return nil if errors.nil?
+
+      errors.is_a?(ActiveModel::Errors) ? errors.messages : errors
+    end
+  end
+end
